The pollfd fields events & revents had the wrong size (int instead of short).
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@1749 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/posix/poll.h
@@ -7,9 +7,9 @@
 typedef unsigned long nfds_t;
 
 struct pollfd {
-	int fd;
-	int events;		/* events to look for */
-	int revents;	/* events that occured */
+	int		fd;
+	short	events;		/* events to look for */
+	short	revents;	/* events that occured */
 };
 
 /* events & revents - compatible with the B_SELECT_xxx definitions in Drivers.h */
